Decomposed submission script into an installable package
