+++ .gitignore
@@ -8,6 +8,7 @@
 .ruff_cache
 .envrc
 *.pyc
+uv.lock
 
 # locally customizable script to push a build of a development version to a s3/http accessible location for testing
 push_dev.local
