+++ README.md
@@ -27,3 +27,49 @@ Nextflow child jobs.
 
 For more information on submission script usage, please see the
 [submission script usage documentation](submit/README.md).
+
+
+
+## Building and using development versions of the nextflow plugin
+
+Build the plugin from the current working tree and push it to a
+location that is accessible via s3 or https from the Fuzzball
+cluster.
+
+```sh
+make push_dev
+```
+
+This depends on a script called `push_dev.local` you can create to customize
+where the plugin is pushed for use with your Fuzzball workflows. The plugin's
+name has to match the pattern the submission script looks for. Here is an
+example:
+
+```sh
+#! /bin/bash
+
+ver="${1:-none}"
+fb_ver="${2:=none}"
+if [[ "${ver}${fb_ver}" =~ none ]] ; then
+    echo "USAGE:  $0 PLUGIN_VERSION FUZZBALL_VERSION"
+    echo "        PLUGIN_VERSION and FUZZBALL_VERSION are both expected to start with a v"
+fi
+
+aws s3 cp \
+    build/distributions/nf-fuzzball-${ver}.zip \
+    s3://MY_BUCKET/nf-fuzzball/${ver}/nf-fuzzball-${ver}-stable-${fb_ver}.zip
+```
+
+Note the naming pattern `${ver}/nf-fuzzball-${ver}-stable-${fb_ver}.zip` where
+`$ver` is the full version of the plugin (e.g. 'v0.1.2') and `$fb_ver` is the
+MAJOR.MINOR version of fuzzball to build against (e.v. 'v2.2'). To use this
+build with the nf-fuzzball-submit script you would specify the location to
+download from like so:
+
+```sh
+python submit_nextflow.py \
+    --plugin-base-uri=s3://MY_BUCKET/MY_PREFIX/nf-fuzzball  \
+    --s3-secret "secret://user/MY_S3_SECRET" \
+    -- \
+    nextflow run hello
+```
